Add .flaskenv and index.py
This commit adds a .flaskenv file to the project root directory. This file
contains environment variables that are used by Flask. The file also contains
the SECRET_KEY variable which is used by Flask-WTF to protect against CSRF
attacks.

This commit also adds an index.py file to the project root directory. This file
is used to create the Flask application instance.